+++ compiler-core/src/analyse/name.rs
@@ -11,12 +11,6 @@ use crate::{
 use super::{Error, Named};
 use heck::{ToSnakeCase, ToUpperCamelCase};
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct NameCorrection {
-    pub location: SrcSpan,
-    pub correction: EcoString,
-}
-
 static VALID_NAME_PATTERN: OnceLock<Regex> = OnceLock::new();
 
 fn valid_name(name: &EcoString) -> bool {
@@ -45,10 +39,13 @@ fn valid_upname(name: &EcoString) -> bool {
 
 pub fn check_name_case(location: SrcSpan, name: &EcoString, kind: Named) -> Result<(), Error> {
     let valid = match kind {
-        Named::Type | Named::TypeVariable | Named::CustomTypeVariant => valid_upname(name),
-        Named::Variable | Named::Argument | Named::Label | Named::Constant | Named::Function => {
-            valid_name(name)
-        }
+        Named::Type | Named::TypeAlias | Named::CustomTypeVariant => valid_upname(name),
+        Named::Variable
+        | Named::TypeVariable
+        | Named::Argument
+        | Named::Label
+        | Named::Constant
+        | Named::Function => valid_name(name),
         Named::Discard => valid_discard_name(name),
     };
 
@@ -63,32 +60,26 @@ pub fn check_name_case(location: SrcSpan, name: &EcoString, kind: Named) -> Resu
     })
 }
 
-pub fn correct_name_case(location: SrcSpan, name: &EcoString, kind: Named) -> NameCorrection {
-    let correction = match kind {
-        Named::Type | Named::TypeVariable | Named::CustomTypeVariant => {
+pub fn correct_name_case(name: &EcoString, kind: Named) -> EcoString {
+    match kind {
+        Named::Type | Named::TypeAlias | Named::CustomTypeVariant => {
             name.to_upper_camel_case().into()
         }
-        Named::Variable | Named::Argument | Named::Label | Named::Constant | Named::Function => {
-            name.to_snake_case().into()
-        }
+        Named::Variable
+        | Named::TypeVariable
+        | Named::Argument
+        | Named::Label
+        | Named::Constant
+        | Named::Function => name.to_snake_case().into(),
         Named::Discard => eco_format!("_{}", name.to_snake_case()),
-    };
-    NameCorrection {
-        location,
-        correction,
     }
 }
 
-pub fn check_argument_names(
-    names: &ArgNames,
-    problems: &mut Problems,
-    name_corrections: &mut Vec<NameCorrection>,
-) {
+pub fn check_argument_names(names: &ArgNames, problems: &mut Problems) {
     match names {
         ArgNames::Discard { name, location } => {
             if let Err(error) = check_name_case(*location, name, Named::Discard) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*location, name, Named::Discard));
             }
         }
         ArgNames::LabelledDiscard {
@@ -99,17 +90,14 @@ pub fn check_argument_names(
         } => {
             if let Err(error) = check_name_case(*label_location, label, Named::Label) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*label_location, label, Named::Label));
             }
             if let Err(error) = check_name_case(*name_location, name, Named::Discard) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*name_location, name, Named::Discard));
             }
         }
         ArgNames::Named { name, location } => {
             if let Err(error) = check_name_case(*location, name, Named::Argument) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*location, name, Named::Argument));
             }
         }
         ArgNames::NamedLabelled {
@@ -120,11 +108,9 @@ pub fn check_argument_names(
         } => {
             if let Err(error) = check_name_case(*label_location, label, Named::Label) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*label_location, label, Named::Label));
             }
             if let Err(error) = check_name_case(*name_location, name, Named::Argument) {
                 problems.error(error);
-                name_corrections.push(correct_name_case(*name_location, name, Named::Argument));
             }
         }
     }